Update app.R
Adding the evaluation of the input data. A list of functions, which is great, it automatice the app. It lets me choose which function to use with less lines of code. Dates also are functional to respect certain patterns to work with, so it'll will tell the user how to input the date format, till I find a better solution.