Update app.R
The application now executes the python script only when it notices that an hour or more have passed since the last update. It also display the 0.00USD strings when you haven't even imported the data. Just to keep the format of the window. The Application also displays the date and time of the last update of the prices.