Update app.R
Ok, I've updated this app way more. Now, this app is able to upload the user files, but only excel files, since it can manage multiple sheets. The app, in principle, will use different sheets because of organization at coding and recording for the user. So, it would mean that the tables would repeat themselves a few times, and also, their functions. So it was a good idea to use a module in it. So It has a functional module that let the app show and operate each table of the user file. Not only that but also create new files.

What is yet to be working on is the download section, the data treatment between USD and other currencies. And other instances. But for now, it got better. The code is being reduced and better functional. Reactive Values are better to be worked with.