Update app.R
Now you can view the tail end of your uploaded file on the initial tab and change the section of the file you want to see. This app is currently functional for recording purposes. However, some adjustments are still needed in the data input section: for instance, automatically filling some lines with USD prices. Additionally, we need to create a tab that summarizes the information in your dataset. This summary might also be integrated into the initial tab. We'll decide as we proceed.