Update app.R
We'll add one more tab where the data will be visually descriptive
I will try to create charts
--> It will be added a vertical tab strip on the left side
There will be 3 tabs. The first one for input where this information is entered
The second tab will be for charts and time series, with descriptive tables.
The third tab is yet to be decided.